requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
Pillow==8.2.0
transformers==4.17.0
torch==2.1.0
torchvision==0.16.0
